+++ resources/assets/js/components/partials/Avatar.vue
@@ -5,15 +5,35 @@
   <div>
     <template v-if="clickable == true">
       <a :href="'people/' + contact.id">
-        <img v-if="contact.avatar_url" v-tooltip.bottom="contact.complete_name" :src="contact.avatar_url" class="br4 h3 w3 dib tc" />
-        <div v-else v-tooltip.bottom="contact.complete_name" :style="{ 'background-color': contact.default_avatar_color }" class="br4 h3 w3 dib pt3 white tc f4">
+        <img v-if="check"
+             v-tooltip.bottom="contact.complete_name"
+             class="br4 h3 w3 dib tc"
+             :alt="contact.initials"
+             :src="contact.avatar_url"
+             @error="check=false"
+        />
+        <div v-else
+             v-tooltip.bottom="contact.complete_name"
+             class="br4 h3 w3 dib tc white f3"
+             :style="'padding-top: 12px; background-color: '+ contact.default_avatar_color"
+        >
           {{ contact.initials }}
         </div>
       </a>
     </template>
     <template v-else>
-      <img v-if="contact.avatar_url" v-tooltip.bottom="contact.complete_name" :src="contact.avatar_url" class="br4 h3 w3 dib tc" />
-      <div v-else v-tooltip.bottom="contact.complete_name" :style="{ 'background-color': contact.default_avatar_color }" class="br4 h3 w3 dib pt3 white tc f4">
+      <img v-if="check"
+           v-tooltip.bottom="contact.complete_name"
+           class="br4 h3 w3 dib tc"
+           :alt="contact.initials"
+           :src="contact.avatar_url"
+           @error="check=false"
+      />
+      <div v-else
+           v-tooltip.bottom="contact.complete_name"
+           class="br4 h3 w3 dib tc white f3"
+           :style="'padding-top: 12px; background-color: '+ contact.default_avatar_color"
+      >
         {{ contact.initials }}
       </div>
     </template>
@@ -32,5 +52,10 @@ export default {
       default: true,
     },
   },
+  data () {
+    return {
+      check: true,
+    };
+  }
 };
 </script>